pyproject.toml
[project]
name = "twitter-bookmarks"
version = "0.2.0"
description = "Scrape X bookmarks via twitter-cli, enrich, cluster with embeddings, and emit a hierarchical Obsidian vault"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "trafilatura>=1.12",
    "openai>=1.50",
    "pydantic>=2.9",
    "fastembed>=0.4",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "twitter-cli>=0.5",
]

[tool.uv]
package = false
